yt-rag/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart

langchain
langchain-core
langchain-google-genai
langchain-pinecone
langgraph

pinecone-client
youtube-transcript-api
